move pageRefObs from discourseGraphMode
